+++ .gitignore
@@ -1,4 +1,14 @@
-HELP.md
+*.class
+*.log
+*.jar
+*.war
+*.nar
+*.ear
+*.zip
+*.tar.gz
+*.rar
+hs_err_pid*
+out/
 target/
 !.mvn/wrapper/maven-wrapper.jar
 !**/src/main/**/target/
@@ -31,3 +41,6 @@ build/
 
 ### VS Code ###
 .vscode/
+
+*.db
+db/*
